test(app): repair stale CI coverage
