+++ apps/app/tests/work-context-entry-wiring.test.ts
@@ -67,8 +67,8 @@ describe("personal and Enterprise chat entry wiring", () => {
   });
 
   test("keeps the selected template library scope when an Enterprise launches a template", () => {
-    expect(sessionPage).toContain("template.manifest.id,\n            templateResourceScope,");
-    expect(sessionPage).toContain("props.selectedSessionId,\n                            undefined,\n                            templateResourceScope,");
+    expect(sessionPage).toMatch(/template\.manifest\.id,\s+templateResourceScope,/);
+    expect(sessionPage).toMatch(/props\.selectedSessionId,\s+undefined,\s+templateResourceScope,/);
     expect(sessionRoute).toContain("templateScope ?? readActiveWorkContextId()");
     expect(sessionRoute).toContain("Template unavailable");
     expect(sessionRoute).toContain("deleteSession(endpoint.workspaceId, createdSessionId)");
